+++ src/Phan/Language/EmptyUnionType.php
@@ -960,7 +960,7 @@ final class EmptyUnionType extends UnionType
      * @return UnionType
      * The subset of types in this
      */
-    public function genericArrayElementTypes() : UnionType
+    public function genericArrayElementTypes(bool $add_real_types = false) : UnionType
     {
         return $this; // empty
     }
@@ -1447,6 +1447,11 @@ final class EmptyUnionType extends UnionType
     }
 
     public function eraseRealTypeSet() : UnionType
+    {
+        return $this;
+    }
+
+    public function eraseRealTypeSetRecursively() : UnionType
     {
         return $this;
     }