+++ tests/files/src/0440_no_warn_set_check.php
@@ -4,6 +4,6 @@ function test(array $data = []) : string {
     if (!isset($data['key'])) {
         $data['key'] = [];
     }
-    return $data['key'];
+    return $data['key'];  // TODO: The inferred real type could be improved (e.g. unknown?)
 }
 test();